// packages/shared/types/interface.d.ts
export interface OpenFileOptions {
  title?: string;
  filters?: { name: string; extensions: string[] }[];
  multiple?: boolean;
}

export type EditField = 'start' | 'end' | 'original' | 'translation';
export type EditArgs = {
  index: number;
  field: EditField;
  value: number | string;
};

export interface SrtSegment {
  index: number;
  start: number;
  end: number;
  original: string;
  translation?: string;
  reviewedInBatch?: number;
}

export interface OpenFileResult {
  filePaths?: string[]; // Made optional as checks exist in usage
  fileContents?: string[]; // Made optional as checks exist in usage
  error?: string;
  canceled?: boolean;
}